golden_hour_model/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0

# Synthetic data generation
sdv>=1.6.0
ctgan>=0.7.0

# Data quality and validation
great-expectations>=0.18.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
